Integrate neural taggers and CLIP-based classifier
