+++ README.md
@@ -37,8 +37,13 @@ cd image-tagbot
 ```
 2. Установите токен бота в переменную окружения `BOT_TOKEN`:
 ```bash
-export BOT_TOKEN="ВАШ_ТОКЕН"
-```
+ export BOT_TOKEN="ВАШ_ТОКЕН"
+ ```
+
+3. (Опционально) укажите путь к базе данных:
+   ```bash
+   export DATABASE_URL="sqlite:///./photos.db"
+   ```
 
 ### Модели тегирования
 
@@ -65,3 +70,11 @@ export BOT_TOKEN="ВАШ_ТОКЕН"
 - `data/docs/`
 
 Убедитесь, что эти каталоги доступны на запись.
+
+### Классификация изображений
+
+Определение типа изображения выполняется моделью CLIP. Настроить её можно
+переменными окружения:
+
+- `CLASSIFIER_MODEL` — архитектура модели (по умолчанию `ViT-B-32`)
+- `CLASSIFIER_MODEL_PATH` — имя предобученных весов или путь к ним (по умолчанию `openai`)
